Lookup participants by email
Adds the ability to search for a participant by email address. If more
than one match is found (should not be possible), the user is redirected
to the participant page. If a single result is found (most common), the
user will see the results page.

Related: https://precisionmedicineinitiative.atlassian.net/browse/HPRO-327