[DASH1-71] Replace date picker in Dashbaords with common from HealthPro
The ticket called for fixing missing display arrows in the jQuery UI
implementation of the date picker, but that is not the one we are using
elsewhere in HealthPro. In the interest of standardizing on a single
implementation, this PR uses the convention from HealthPro. It adds
some additional controls around visually restricting "linked" date
fields from picking a date out of range from the set start/end window.

More generally speaking, the JavaScript needs some :heart:, but this
resolves the pressing matter.

[[DASH1-71](https://precisionmedicineinitiative.atlassian.net/browse/DASH1-71)]