[DASH1-75] Fix 100 day bug in Realtime tab
Prior to this PR, a date range greater than 100 days would cause the realtime tab to fail because of an uncaught API error. This PR adjusts the logic so that a realtime (`history=FALSE`) request will batch in 100 day totals, while `history=TRUE` can handle 600 day increments. Because Memcache is capped at 1 MiB per key, it is reduced further to 200.

[[DASH1-75](https://precisionmedicineinitiative.atlassian.net/browse/DASH1-75)]