[DASH1-101] Dashboard > Update recruitment center GPS coordinates
Two of the rows in `recruitment_center_codes` have an incorrect longitude (should be negative) which causes the map view to place the points in the upper left corner. This PR introduces a migration to update those rows.

It also includes a trivial change to convert tabs to spaces in the original migration, but does not change its content.

[[DASH1-101](https://precisionmedicineinitiative.atlassian.net/projects/DASH1/issues/DASH1-101)]