[DASH1-140] Fix filter toggles where name has parenthesis
The prior implementation of this took the name, lowercased it, split on
whitespace and then joined it back together with dashes. Trouble with
that is when the name contains other characters that may trip up the
selector in jQuery. This implementation abstracts that away into a
custom Twig function that 'slugify's the the name a bit better, though
there are still cases where an org name could trip it up. If we needed a
more robust solution, I'd advise using a third-party package specialized
to the task.

[[DASH1-140](https://precisionmedicineinitiative.atlassian.net/browse/DASH1-140)]