[DASH1-141] Refactor EHR dashboard to use single endpoint
This is the follow-on work to DASH1-124. During QA, we caught some inconsistencies in the data being returned between the ParticipantsOverTime and Organizations endpoint, most likely attributed to cache timing. There is a consolidated endpoint (by omitting the type of data the user is after) that would guarantee the two data sets would be in sync at all times. This ticket is to refactor the existing dashboard to rely on that approach.

[[DASH1-141](https://precisionmedicineinitiative.atlassian.net/browse/DASH1-141)]