[HPRO-683] Deceased Reports > Add badge to Welcome Page
Adds a badge that shows the number of Pending Deceased Reports for the logged in user on the main Welcome page.

Note: It takes a second or so for it to show up, and will only appear if the organization has pending reports (i.e. it won't appear if there are
zero pending reports).

[HPRO-683]