[HPRO-687] Disable deceased reporting for withdrawn/deactivated participants
This PR removes the cog icon on the Participant Summary page for
deactivated and withdrawn participants. It also adds logic that prevents
new Deceased Reports from being created for those participants.

Additionally, it fixes a bug where the Deactivated filter would not work
in the Work Queue because of a restriction on combining it with Deceased
Status.

[HPRO-687]