[HPRO-699] Update README for credential scanning
During our conversion to the second generation runtime for App Engine,
we lost the composer `post-install-cmd` stanzy because it prevented
deployments (doesn't mesh well with their Cloud Deploy tool). As a
result, we no longer have a way of ensuring that a developer has
installed the `git-secrets` utiliity that scans for potentially leaked
credentials prior to commit. This PR updates the README to include
instructions to run the helper script manually.

[HPRO-699]